refactor: 테스트 및 컨트롤러 리팩토링
- deletePost_InvalidPermissionError_DirectCall 테스트 추가
- PostController 권한 어노테이션 추가 적용, 상태 코드 변경
- 댓글 작성 API에 유효성 검사 추가 (request.getContent() 체크)
- PostEntityFixture 활용, 중복코드 제거
- PostEntity의 likePost(), unlikePost() 반환값 수정 (성공여부 반환)

+++ src/main/java/com/seeyoungryu/connecti/controller/PostController.java
@@ -18,6 +18,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.Authentication;
@@ -70,12 +71,16 @@ public class PostController {
     /*
     post 삭제
      */
+    //Spring Security , Enum : ROLE_USER -> Todo : @Secured("ROLE_USER")를 메서드 단위가 아닌 서비스 계층에서 적용하는 것이 더 유연함? 수정필요?
     @DeleteMapping("/{postId}")
     @Secured("ROLE_USER")
-    //Spring Security , Enum : ROLE_USER -> Todo : @Secured("ROLE_USER")를 메서드 단위가 아닌 서비스 계층에서 적용하는 것이 더 유연함? 수정필요?
-    public ApiResponse<Void> deletePost(@PathVariable Long postId, Authentication authentication) {
+    public ResponseEntity<ApiResponse<Void>> deletePost(@PathVariable Long postId, Authentication authentication) {
+        if (authentication == null || authentication.getName() == null) {
+            throw new ConnectiApplicationException(ErrorCode.INVALID_PERMISSION);
+        }
+
         postService.deletePost(authentication.getName(), postId);
-        return ApiResponse.success(null);
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).body(ApiResponse.success(null)); // 상태 코드 수정 (204 No Content)
     }
